Use template file instead of inline code
